+++ src/main/java/com/jonathan/futebol_api/adapter/dto/ClubeResponseDTO.java
@@ -3,7 +3,7 @@ package com.jonathan.futebol_api.adapter.dto;
 import jakarta.validation.constraints.NotBlank;
 
 public record ClubeResponseDTO(
-        Integer id,
+        Long id,
         @NotBlank String nome,
         String estado,
         String estadio,
